Finally got everything setup and tests working!

+++ AsunaSpring/src/main/java/com/grantcallant/asunaspring/repository/eso/model/Parse.java
@@ -29,7 +29,7 @@ public class Parse
   private EsoCharacter esoCharacter;
 
   @Column(name = "dps", columnDefinition = "INT UNSIGNED not null")
-  private Long dps;
+  private int dps;
 
   @Column(name = "parse_file_key", nullable = false)
   private UUID parseFileKey;
